Added avro parsing in spark (à tester)

+++ spark/connexion/src/main/scala/TicketAccumulator.scala
@@ -3,22 +3,22 @@ package Infra
 import org.apache.spark.util.AccumulatorV2
 
 
-case class TicketAccumulator(initialValue : Long = 0) extends AccumulatorV2[Long, Long] {
+case class TicketAccumulator(initialValue : Float = 0) extends AccumulatorV2[Float, Float] {
 
 	private var _value = initialValue
 
-	override def value:Long = _value
+	override def value:Float = _value
 
 	override def isZero: Boolean = value == 0
 
-	override def copy(): AccumulatorV2[Long, Long] = TicketAccumulator(value)
+	override def copy(): AccumulatorV2[Float, Float] = TicketAccumulator(value)
 
 	override def reset(): Unit = value
 
-	override def add(v: Long): Unit = {
+	override def add(v: Float): Unit = {
 		_value += v
 	}
 
-	override def merge(other: AccumulatorV2[Long, Long]): Unit = add(other.value)
+	override def merge(other: AccumulatorV2[Float, Float]): Unit = add(other.value)
 
 }
